requirements.txt
# FastAPI and Pydantic for API development and validation
fastapi==0.100.0
pydantic==1.10.2

# Uvicorn for ASGI server
uvicorn==0.22.0

# sumy for text summarization (LexRank)
sumy==0.8.1

# YouTube transcript API for extracting transcripts from YouTube videos
youtube-transcript-api==0.5.0

# Regular expressions for string processing
regex==2024.10.9  # regex is often already included with Python, but you can ensure it's included

# PyMongo for MongoDB interactions
pymongo==4.7.0

# Python-dotenv for environment variable loading
python-dotenv==1.0.0

# bson for MongoDB ObjectId (comes with PyMongo but added for completeness)
bson==0.5.10
